refactor: API object 파일로 분리
