--- app/src/main/java/com/example/mvvm2/MyApplication.kt
@@ -1,25 +1,0 @@
-package com.example.mvvm2
-
-import android.app.Application
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-class MyApplication : Application() {
-    companion object {
-        const val BASE_URL = "https://openapi.naver.com/v1/search/"
-
-        private val client = OkHttpClient.Builder()
-            .readTimeout(5000, TimeUnit.MILLISECONDS)
-            .connectTimeout(5000, TimeUnit.MILLISECONDS)
-            .addNetworkInterceptor(XHeaderInterceptor()) // 자동 헤더 전송
-            .build()
-
-        val retrofit: Retrofit = Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .client(client)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-}
