+++ app/src/main/java/com/example/mvvm2/viewmodel/MovieListViewModel.kt
@@ -2,8 +2,7 @@ package com.example.mvvm2.viewmodel
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.mvvm2.MyApplication
-import com.example.mvvm2.api.SearchService
+import com.example.mvvm2.api.Api
 import com.example.mvvm2.model.Movie
 import com.example.mvvm2.model.SearchResponse
 import retrofit2.Call
@@ -15,14 +14,13 @@ class MovieListViewModel : ViewModel() {
 
     fun searchMovie(query: String) {
 
-        val searchService = MyApplication.retrofit.create(SearchService::class.java)
-        searchService.searchMovie(query).enqueue(object : Callback<SearchResponse> {
+        Api.searchApi.searchMovie(query).enqueue(object : Callback<SearchResponse> {
             override fun onResponse(
                 call: Call<SearchResponse>,
                 response: Response<SearchResponse>
             ) {
                 val searchResponse = response.body()
-                movieList.value = searchResponse!!.movies
+                movieList.value = searchResponse?.movies
             }
 
             override fun onFailure(call: Call<SearchResponse>, t: Throwable) {
